+++ typescript/utils/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @hyperlane-xyz/utils
 
+## 13.2.0
+
 ## 13.1.1
 
 ### Patch Changes